react-agent-function-calling/pyproject.toml
[project]
name = "react-agent-function-calling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain>=1.2.0",
    "langchain-google-genai>=4.1.1",
    "langchain-tavily>=0.2.15",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
]
